refactor: move arg logic to ape.cli
